Add flag to popper run to ignore undefined secrets
When travis jobs run PRs from forks of the main repo, they fail if the
secrets defined in the popper workflow are not defined, as this is an
outcome of the way Travis deals with secrets (forks cannot access
secrets for the main repo they were forked from).

Popper now provides a flag that allows the execution of a workflow to
continue (without erroring out) when a secret is not defined. It's up to
the workflow to consider the case when the corresponding environment
variable hasn't been defined. In the case of the workflow for this repo,
we only access these variables on the last step (the one that pushes the
docker image to docker), so it is fine if these are not defined when
running on branches other than master.